adding new download method, unzip dataset for train, tram and bus. adding method to load dataset by mode(train, tram, bus), adding method to load combine all dataset
