+++ artificial-intelligence/chatbot/code/mpt_bot/credentials.yml
@@ -21,10 +21,13 @@ rest:
 #  user_message_evt: <event name for user message>
 #  bot_message_evt: <event name for bot messages>
 #  session_persistence: <true/false>
+#socketio:
+#    user_message_evt: user_uttered
+#    bot_message_evt: bot_uttered
 socketio:
-    user_message_evt: user_uttered
-    bot_message_evt: bot_uttered
-
+  user_message_evt: user_uttered
+  bot_message_evt: bot_uttered
+  session_persistence: true
 #mattermost:
 #  url: "https://<mattermost instance>/api/v4"
 #  token: "<bot token>"
