+++ artificial-intelligence/chatbot/code/mpt_bot/data/rules.yml
@@ -71,11 +71,6 @@ rules:
   - intent: ask_quickest_route
   - action: action_find_best_route_with_transfers
 
-- rule: Find next train timing
-  steps:
-  - intent: ask_next_train
-  - action: action_find_next_train
-
 - rule: Ask for location when running direction script
   steps:
   - intent: run_direction_script
@@ -90,3 +85,18 @@ rules:
   steps:
   - intent: provide_destination
   - action: action_run_direction_script
+
+- rule: Find next train timing
+  steps:
+  - intent: ask_next_train
+  - action: action_find_next_train
+  
+- rule: Respond to next bus query
+  steps:
+  - intent: find_next_bus
+  - action: action_find_next_bus
+
+- rule: Respond to next tram query
+  steps:
+  - intent: find_next_tram
+  - action: action_find_next_tram
