+++ artificial-intelligence/chatbot/code/mpt_bot/domain.yml
@@ -8,6 +8,8 @@ intents:
   - greet
   - goodbye
   - find_next_train
+  - find_next_bus
+  - find_next_tram
   - request_map
   - find_best_route
   - ask_direct_route
@@ -131,6 +133,8 @@ responses:
   - text: "Great! Now, please provide your destination."
 
 actions:
+- action_find_next_tram
+- action_find_next_bus
 - action_find_next_train
 - action_generate_map
 - action_find_best_route
