+++ artificial-intelligence/chatbot/code/mpt_bot/index.html
@@ -55,7 +55,48 @@
   </head>
 
   <body>
-    <div class="container">
+    <div class="container">      
+		<table border="1" cellpadding="5" cellspacing="0" style="width: 100%;">
+			<tr>
+				<th colspan="2" style="font-weight:bold;text-align:left;">Train stations map</th>
+			</tr>
+			<tr>
+				<td colspan="2">Show me a map of Melbourne train stations?</td>
+			</tr>
+			<tr>
+				<th colspan="2" style="font-weight:bold;text-align:left;">Route planning</th>
+			</tr>
+			<tr>
+				<td>What’s the best way to go from Box Hill to Glenferrie?</td>
+			</tr>
+			<tr>
+				<td>Which train should I take to get from Camberwell to Canterbury?</td>
+			</tr>
+			<tr>
+				<th colspan="2" style="font-weight:bold;text-align:left;">Schedule - timing information</th>
+			</tr>
+			<tr>
+				<td>When is the next train from Ringwood to Parliament station?</td>
+			</tr>
+			<tr>
+				<th colspan="2" style="font-weight:bold;text-align:left;">Transfers and connection</th>
+			</tr>
+			<tr>
+				<td>How do I get from Flinders Street to Melbourne Central?</td>
+			</tr>
+			<tr>
+				<td>How many transfers are there between North Melbourne to Hawthorns?</td>
+			</tr>
+			<tr>
+				<th colspan="2" style="font-weight:bold;text-align:left;">Route optimisation</th>
+			</tr>
+			<tr>
+				<td>Which route has the fewest stops between Ringwood and Parliament station?</td>
+			</tr>
+			<tr>
+				<td>What is the route with the least number of stops from Dandenong station to Parliament station?</td>
+			</tr>
+		</table>		
       <!-- Modal for rendering the charts, declare this if you want to render charts, 
          else you remove the modal -->
       <div id="modal1" class="modal">
@@ -66,7 +107,7 @@
       <div class="widget">
         <div class="chat_header">
           <!--Add the name of the bot here -->
-          <span class="chat_header_title">Alex</span>
+          <span class="chat_header_title">MPT Chatbot</span>
           <span class="dropdown-trigger" href="#" data-target="dropdown1">
             <i class="material-icons"> more_vert </i>
           </span>
